% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLociStatTimeCourse.R
\name{calcLociStatTimeCourse}
\alias{calcLociStatTimeCourse}
\title{Calculate the per-cytosine statistics for time-course data.}
\usage{
calcLociStatTimeCourse(bs.object, meta, force.slope = FALSE, mc.cores = 1)
}
\arguments{
\item{bs.object}{a \code{BSseq} object from the \code{bsseq} package.}

\item{meta}{a \code{data.frame}.}

\item{force.slope}{if \code{TRUE}, we force the slope in the linear
model to be the same between two conditions. Otherwise, the slopes are
fitted separately but not tested.}

\item{mc.cores}{number of cores used for the parallelization.}
}
\value{
A \code{MethCP} object that is not segmented.
}
\description{
For each cytosine, \code{calcLociStatTimeCourse} fits a linear model
on the arcsin-tranformed methylation ratios, and test the differences
of the slope between the treatment and the control group.
}
\details{
\code{bs.object} is a \code{BSseq} object from the \code{bsseq} package,
which contains the raw data including coverges, methylated counts and
position infomation for every cytosine in the dataset.
\code{meta} must contain columns named \code{Condition}, \code{Time}
and \code{SampleName} in the dataframe. They are used to fit the linear
model.
}
\examples{

}
