% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCP.R
\name{getSigRegion}
\alias{getSigRegion}
\title{Obtain the significant DMRs.}
\usage{
getSigRegion(object, sig.level = 0.01, mean.coverage = 1, mean.diff = 0.1,
  nC.valid = 10)
}
\arguments{
\item{object}{a \code{MethCP} object that is segmented using function
\code{segmentMethCP}.}

\item{sig.level}{significance level to call a region DMR.}

\item{mean.coverage}{The minimum average coverage required for the
reported DMRs.}

\item{mean.diff}{The minimum differences between groups required for
the reported DMRs.}

\item{nC.valid}{number of valid cytosines required for the reported DMRs.}
}
\value{
a \code{data.frame} containing the DMRs.
}
\description{
\code{getSigRegion} returns the significant DMRs giving the segmented
\code{MethCP} object.
}
\examples{


}
