% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCP.R
\docType{class}
\name{MethCP-class}
\alias{MethCP-class}
\title{Class \code{MethCP}}
\description{
A class for performing DMR (differentially methylated region) detection
analysis using method \code{MethCP} on whole genome bisulfite sequencing
data.
}
\section{Slots}{

\describe{
\item{\code{test}}{a character string of the name of the per-cytosine statistcis.}

\item{\code{group1}}{a character vector containing the sample names of the treatment
group.}

\item{\code{group2}}{a character vector containing the sample names of the control
group.}

\item{\code{stat}}{a \code{GRangesList} object containing the results of
per-cytosine tests.}

\item{\code{segmentation}}{a \code{GRanges} object containing the segments and their
infomation such as region-based statistics, coverages, etc.}
}}

\examples{

}
