% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCP.R
\name{MethCP}
\alias{MethCP}
\title{The constructor function for \code{MethCP} objects.}
\usage{
MethCP(test = NA_character_, group1 = NA, group2 = NA, chr, pos, pvals,
  effect.size, ...)
}
\arguments{
\item{test}{a character string of the name of the per-cytosine statistcis.}

\item{group1}{a character vector containing the sample names of the treatment
group.}

\item{group2}{a character vector containing the sample names of the control
group.}

\item{chr}{a character vector containing the cytosine chromosome infomation.}

\item{pos}{a numeric vector containing the cytosine positions.}

\item{pvals}{a numeric vector containing the \code{p}-values for each cytosine.}

\item{effect.size}{a numeric vector containing the effect sizes for each cytosine.}

\item{testName}{a character string containing the names of the test to be
performed per cytosine.}
}
\description{
The constructor function for \code{MethCP} objects.
}
\details{
If not specified by function \code{calcLociStatTimeCourse},
\code{calcLociStat}, the parameter \code{test} can be set to any user-specified
string indicating the name of the test performed.

In the cases where the goal is not to compare between treatment and control
groups, parameter \code{group1} and \code{group2} can be set to \code{NA}.

If generated by \code{calcLociStat}, parameter \code{stat} will be a
\code{GRangesList} object where each element in the list contains statistics
for each of the chromosome in the dataset.
}
\examples{
obj <- MethCP(test = "myTest",
              group1 = paste0("Treatment", 1:3),
              group2 = paste0("Control", 1:3),
              chr = rep("Chr01", 5),
              pvals = c(0, 0.1, 0.9, NA, 0.02),
              pos = c(2, 5, 9, 10, 18),
              effect.size = c(1, -1, NA, 9, Inf))
# MethCP will omit the NAs and infinite values.
obj

}
