% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCP.R
\name{show,MethCP-method}
\alias{show,MethCP-method}
\title{The show method}
\usage{
\S4method{show}{MethCP}(object)
}
\arguments{
\item{object}{a \code{MethCP} object.}
}
\value{
No value will be returned.
}
\description{
Print \code{MethCP} object information.
}
