% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethCP.R
\name{segmentMethCP}
\alias{segmentMethCP}
\title{Perform segmentation on a \code{MethCP} object.}
\usage{
segmentMethCP(methcp.object, bs.object, region.test = c("fisher", "stouffer",
  "weighted-variance", "weighted-coverage"), mc.cores = 1, min.width = 2,
  sig.level = 0.01, presegment_dist = 600, ...)
}
\arguments{
\item{methcp.object}{a \code{MethCP} object.}

\item{bs.object}{a \code{BSseq} object from the \code{bsseq} package.}

\item{region.test}{The meta-analysis method used to create region-based
test statistics.}

\item{mc.cores}{number of cores used for the parallelization.}

\item{min.width}{the minimum width for the segments, which is used as
termination rule for the segmentation algorithm.}

\item{sig.level}{the significance level of the segments, which is used as
termination rule for the segmentation algorithm.}

\item{presegment_dist}{the maximum distance between cytosines for the
presegmentation.}
}
\value{
a \code{MethCP} object that is not segmented.
}
\description{
Perform CBS algorithm that segments the genome into
similar levels of sigficance.
}
\details{
The \code{MethCP} object \code{methcp.object} can be generated from
functions \code{calcLociStat}, \code{calcLociStatTimeCourse}, or
\code{methcpFromStat}.

If \code{region.test = "fisher"}, Fisher's combined probability test is used.

If \code{region.test = stouffer} Stouffer's test is applied.

If \code{region.test = "weighted-variance"} we use the variance of the test to
combine per-cytosine based statistcis into a region-based statistic.

If \code{region.test = "weighted-coverage"} we use the coverage of the test to
combine per-cytosine based statistcis into a region-based statistic.
}
\examples{



}
